backend/requirements.txt
fastapi==0.135.3
httpx==0.28.1
PyJWT==2.12.1
pydantic-settings==2.13.1
uvicorn[standard]==0.42.0
